Update interfaces submodule to 4b95458 Add silkinterfaces library for gRPC generated sources Move gRPC generated sources to interfaces folder Refactor build of gRPC proto files
