+++ silkrpc/ethdb/kv/awaitables.hpp
@@ -42,7 +42,7 @@
 #include <silkrpc/ethdb/kv/async_start.hpp>
 #include <silkrpc/ethdb/kv/streaming_client.hpp>
 #include <silkrpc/ethdb/kv/util.hpp>
-#include <silkrpc/ethdb/kv/remote/kv.grpc.pb.h>
+#include <silkrpc/interfaces/remote/kv.grpc.pb.h>
 
 namespace silkrpc::ethdb::kv {
